moved/removed old, unused, and unnecessary files
