+++ KN1DPy/kn1d.py
@@ -4,11 +4,10 @@ import os
 from scipy import interpolate
 
 from .read import sav_read, nc_read
-from .edit_keys import edit_keys
 from .create_shifted_maxwellian import create_shifted_maxwellian
 from .integ_bl import integ_bl
 from .make_dvr_dvx import make_dvr_dvx
-from .sval import sval
+from .utils import sval
 from .interp_fvrvxx import interp_fvrvxx
 from .kinetic_mesh import create_kinetic_h_mesh, create_kinetic_h2_mesh
 from .kinetic_h import kinetic_h
