--- KN1DPy/sign.py
@@ -1,7 +1,0 @@
-#returns the sign of a number
-#   negative => -1
-#   zero => 0
-#   positive => 1
-
-def sign(x):
-  return int(x>0)-int(x<0)
