--- KN1DPy/sval.py
@@ -1,5 +1,0 @@
-# removes leading / trailing spaces
-# truncates string to a specified length
-
-def sval(s,length=None):
-  return str(s).strip()[:length]
