+++ old_or_unused/printcr.py
